0509
叶流就差5件装备没写
SHPC基本已经推完了，还有问题再指

+++ Weapons/SHPC/Effects/BPrePlantera/BossSoulofMightEffect.cs
@@ -22,18 +22,28 @@ namespace CalamityLegendsComeBack.Weapons.SHPC.Effects.BPrePlantera
         public override Color StartColor => new Color(150, 190, 255);  // 高亮（发光边缘）
         public override Color EndColor => new Color(20, 40, 120);      // 深蓝（阴影/核心）
 
-        public override float SquishyLightParticleFactor => 0.01f;
-        public override float ExplosionPulseFactor => 0.01f;
+        public override float SquishyLightParticleFactor => 0f;
+        public override float ExplosionPulseFactor => 0f;
+        public override float GlowScaleFactor => 0f;
+        public override float GlowIntensityFactor => 0f;
         public override bool EnableDefaultSlowdown => false;
 
         public override void OnSpawn(Projectile projectile, Player owner)
         {
-            projectile.Kill();
+            projectile.GetGlobalProjectile<BossSoulofMight_GP>().firstFrame = true;
+            projectile.timeLeft = 2;
+            projectile.penetrate = -1;
+            projectile.tileCollide = false;
         }
 
         public override void AI(Projectile projectile, Player owner)
         {
+            BossSoulofMight_GP gp = projectile.GetGlobalProjectile<BossSoulofMight_GP>();
+            if (!gp.firstFrame)
+                return;
 
+            gp.firstFrame = false;
+            projectile.Kill();
         }
 
         public override void ModifyHitNPC(Projectile projectile, Player owner, NPC target, ref NPC.HitModifiers modifiers)
@@ -86,5 +96,11 @@ namespace CalamityLegendsComeBack.Weapons.SHPC.Effects.BPrePlantera
 
 
 
+    }
+
+    internal class BossSoulofMight_GP : GlobalProjectile
+    {
+        public override bool InstancePerEntity => true;
+        public bool firstFrame;
     }
 }
